修复数字为浮点数

+++ src/utils/toolFunc.js
@@ -9,6 +9,6 @@ export const msgId = createMsgId()
 
 // 判断一个数是否为正整数
 export function isInteger(number){
-    let obj = parseInt(number)
+    let obj = parseFloat(number)
     return  typeof obj=== 'number'&& obj!=0 && obj%1===0
 }
